split get_api_key out as separate fn
